Rename hook to hooks and update imports
